requirements.txt
nose
nose-testconfig
nose_parameterized
selenium
loguru